+++ include/table.h
@@ -1,0 +1,7 @@
+#ifndef TABLE_H
+#define TABLE_H
+
+#include "tetris.h"
+void updateTable(const Piece piece, char (*tablePtr)[ROWS][COLS]);
+
+#endif
