+++ src/controll.c
@@ -1,51 +1,12 @@
 #include "tetris.h"
 
-void spawnNewPiece(Piece *piecePtr) // return s rondom piece
+void spawnNewPiece(Piece *piecePtr)
 {
 	Piece new = copyPiece(getRandomPiece());
 	new.col = rand() % (COLS - new.width + 1);
 	new.row = 0;
 	deletePiece(*piecePtr);
 	*piecePtr = new;
-}
-
-static void fixPieceToTable(const Piece piece, char (*tablePtr)[ROWS][COLS])
-{
-	for (int i = 0; i < piece.width; i++)
-		for (int j = 0; j < piece.width; j++)
-			if (piece.array[i][j])
-				(*tablePtr)[piece.row + i][piece.col + j] = piece.array[i][j];
-}
-
-static int isLineFilled(const int row, const char table[ROWS][COLS])
-{
-	for (int i = 0; i < COLS; i++)
-		if (table[row][i] == 0)
-			return FALSE;
-	return TRUE;
-}
-
-static void putDownLines(const int row, char (*tablePtr)[ROWS][COLS])
-{
-	for (int i = row; i >= 1; i--)
-		for (int j = 0; j < COLS; j++)
-			(*tablePtr)[i][j] = (*tablePtr)[i - 1][j];
-	for (int j = 0; j < COLS; j++)
-		(*tablePtr)[0][j] = 0;
-}
-
-static void checkLines(char (*tablePtr)[ROWS][COLS])
-{
-	int i;
-	for (i = 0; i < ROWS; i++)
-	{
-		if (isLineFilled(i, *tablePtr))
-		{
-			addScore();
-			putDownLines(i, tablePtr);
-			reduceInterval();
-		}
-	}
 }
 
 static void moveDown(Piece *currentPtr, char (*tablePtr)[ROWS][COLS])
@@ -56,8 +17,7 @@ static void moveDown(Piece *currentPtr, char (*tablePtr)[ROWS][COLS])
 		currentPtr->row++;
 	else
 	{
-		fixPieceToTable(*currentPtr, tablePtr);
-		checkLines(tablePtr);
+		updateTable(*currentPtr, tablePtr);
 		spawnNewPiece(currentPtr);
 	}
 	deletePiece(tmp);
@@ -103,5 +63,4 @@ void controllCurrent(Piece *currentPtr, char (*tablePtr)[ROWS][COLS], const int 
 		moveRotate(currentPtr, *tablePtr);
 		break;
 	}
-	printTable(*currentPtr, *tablePtr, getScore());
 }
